Move anonymous-related actions to RefactorProcessor
