+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/corrections/RefactorProcessor.java
@@ -146,12 +146,14 @@ import org.eclipse.jdt.ls.core.internal.corrections.proposals.ChangeCorrectionPr
 import org.eclipse.jdt.ls.core.internal.preferences.PreferenceManager;
 import org.eclipse.jdt.ls.core.internal.text.correction.ActionMessages;
 import org.eclipse.jdt.ls.core.internal.text.correction.RefactorProposalUtility;
+import org.eclipse.jdt.ls.core.internal.text.correction.RefactoringCorrectionCommandProposal;
 import org.eclipse.lsp4j.CodeActionParams;
 
 /**
  * RefactorProcessor
  */
 public class RefactorProcessor {
+	public static final String CONVERT_ANONYMOUS_CLASS_TO_NESTED_COMMAND = "convertAnonymousClassToNestedCommand";
 
 	private PreferenceManager preferenceManager;
 
@@ -176,6 +178,9 @@ public class RefactorProcessor {
 				getExtractMethodProposal(params, context, coveringNode, problemsAtLocation, proposals);
 				getExtractFieldProposal(params, context, problemsAtLocation, proposals);
 				getInlineProposal(context, coveringNode, proposals);
+				getConvertAnonymousToNestedProposals(params, context, coveringNode, proposals);
+				getConvertAnonymousClassCreationsToLambdaProposals(context, coveringNode, proposals);
+				getConvertLambdaToAnonymousClassCreationsProposals(context, coveringNode, proposals);
 			}
 			return proposals;
 		}
@@ -368,6 +373,145 @@ public class RefactorProcessor {
 		}
 
 		return false;
+	}
+
+
+	private boolean getConvertAnonymousToNestedProposals(CodeActionParams params, IInvocationContext context, final ASTNode node, Collection<ChangeCorrectionProposal> proposals) throws CoreException {
+		if (!(node instanceof Name)) {
+			return false;
+		}
+
+		if (proposals == null) {
+			return false;
+		}
+
+		RefactoringCorrectionProposal proposal = null;
+		if (this.preferenceManager.getClientPreferences().isAdvancedExtractRefactoringSupported()) {
+			proposal = getConvertAnonymousToNestedProposal(params, context, node, true /*returnAsCommand*/);
+		} else {
+			proposal = getConvertAnonymousToNestedProposal(params, context, node, false /*returnAsCommand*/);
+		}
+
+		if (proposal == null) {
+			return false;
+		}
+
+		proposals.add(proposal);
+		return true;
+	}
+
+	public static RefactoringCorrectionProposal getConvertAnonymousToNestedProposal(CodeActionParams params, IInvocationContext context, final ASTNode node, boolean returnAsCommand) throws CoreException {
+		String label = CorrectionMessages.QuickAssistProcessor_convert_anonym_to_nested;
+		ASTNode normalized = ASTNodes.getNormalizedNode(node);
+		if (normalized.getLocationInParent() != ClassInstanceCreation.TYPE_PROPERTY) {
+			return null;
+		}
+
+		final AnonymousClassDeclaration anonymTypeDecl = ((ClassInstanceCreation) normalized.getParent()).getAnonymousClassDeclaration();
+		if (anonymTypeDecl == null || anonymTypeDecl.resolveBinding() == null) {
+			return null;
+		}
+
+		final ConvertAnonymousToNestedRefactoring refactoring = new ConvertAnonymousToNestedRefactoring(anonymTypeDecl);
+		if (!refactoring.checkInitialConditions(new NullProgressMonitor()).isOK()) {
+			return null;
+		}
+
+		if (returnAsCommand) {
+			return new RefactoringCorrectionCommandProposal(label, CodeActionKind.Refactor, context.getCompilationUnit(), IProposalRelevance.CONVERT_ANONYMOUS_TO_NESTED, RefactorProposalUtility.APPLY_REFACTORING_COMMAND_ID,
+					Arrays.asList(CONVERT_ANONYMOUS_CLASS_TO_NESTED_COMMAND, params));
+		}
+
+		String extTypeName = ASTNodes.getSimpleNameIdentifier((Name) node);
+		ITypeBinding anonymTypeBinding = anonymTypeDecl.resolveBinding();
+		String className;
+		if (anonymTypeBinding.getInterfaces().length == 0) {
+			className = Messages.format(CorrectionMessages.QuickAssistProcessor_name_extension_from_interface, extTypeName);
+		} else {
+			className = Messages.format(CorrectionMessages.QuickAssistProcessor_name_extension_from_class, extTypeName);
+		}
+		String[][] existingTypes = ((IType) anonymTypeBinding.getJavaElement()).resolveType(className);
+		int i = 1;
+		while (existingTypes != null) {
+			i++;
+			existingTypes = ((IType) anonymTypeBinding.getJavaElement()).resolveType(className + i);
+		}
+		refactoring.setClassName(i == 1 ? className : className + i);
+
+		LinkedProposalModelCore linkedProposalModel = new LinkedProposalModelCore();
+		refactoring.setLinkedProposalModel(linkedProposalModel);
+
+		final ICompilationUnit cu = context.getCompilationUnit();
+		RefactoringCorrectionProposal proposal = new RefactoringCorrectionProposal(label, CodeActionKind.Refactor, cu, refactoring, IProposalRelevance.CONVERT_ANONYMOUS_TO_NESTED);
+		proposal.setLinkedProposalModel(linkedProposalModel);
+		return proposal;
+
+	}
+
+	private static boolean getConvertAnonymousClassCreationsToLambdaProposals(IInvocationContext context, ASTNode covering, Collection<ChangeCorrectionProposal> resultingCollections) {
+		while (covering instanceof Name || covering instanceof Type || covering instanceof Dimension || covering.getParent() instanceof MethodDeclaration
+				|| covering.getLocationInParent() == AnonymousClassDeclaration.BODY_DECLARATIONS_PROPERTY) {
+			covering = covering.getParent();
+		}
+
+		ClassInstanceCreation cic;
+		if (covering instanceof ClassInstanceCreation) {
+			cic = (ClassInstanceCreation) covering;
+		} else if (covering.getLocationInParent() == ClassInstanceCreation.ANONYMOUS_CLASS_DECLARATION_PROPERTY) {
+			cic = (ClassInstanceCreation) covering.getParent();
+		} else if (covering instanceof Name) {
+			ASTNode normalized = ASTNodes.getNormalizedNode(covering);
+			if (normalized.getLocationInParent() != ClassInstanceCreation.TYPE_PROPERTY) {
+				return false;
+			}
+			cic = (ClassInstanceCreation) normalized.getParent();
+		} else {
+			return false;
+		}
+
+		IProposableFix fix = LambdaExpressionsFixCore.createConvertToLambdaFix(cic);
+		if (fix == null) {
+			return false;
+		}
+
+		if (resultingCollections == null) {
+			return true;
+		}
+
+		Map<String, String> options = new HashMap<>();
+		options.put(CleanUpConstants.CONVERT_FUNCTIONAL_INTERFACES, CleanUpOptionsCore.TRUE);
+		options.put(CleanUpConstants.USE_LAMBDA, CleanUpOptionsCore.TRUE);
+		FixCorrectionProposal proposal = new FixCorrectionProposal(fix, new LambdaExpressionsCleanUpCore(options), IProposalRelevance.CONVERT_TO_LAMBDA_EXPRESSION, context, CodeActionKind.Refactor);
+		resultingCollections.add(proposal);
+		return true;
+	}
+
+	private static boolean getConvertLambdaToAnonymousClassCreationsProposals(IInvocationContext context, ASTNode covering, Collection<ChangeCorrectionProposal> resultingCollections) {
+		if (resultingCollections == null) {
+			return true;
+		}
+
+		LambdaExpression lambda;
+		if (covering instanceof LambdaExpression) {
+			lambda = (LambdaExpression) covering;
+		} else if (covering.getLocationInParent() == LambdaExpression.BODY_PROPERTY) {
+			lambda = (LambdaExpression) covering.getParent();
+		} else {
+			return false;
+		}
+
+		IProposableFix fix = LambdaExpressionsFixCore.createConvertToAnonymousClassCreationsFix(lambda);
+		if (fix == null) {
+			return false;
+		}
+
+		// add correction proposal
+		Map<String, String> options = new HashMap<>();
+		options.put(CleanUpConstants.CONVERT_FUNCTIONAL_INTERFACES, CleanUpOptionsCore.TRUE);
+		options.put(CleanUpConstants.USE_ANONYMOUS_CLASS_CREATION, CleanUpOptionsCore.TRUE);
+		FixCorrectionProposal proposal = new FixCorrectionProposal(fix, new LambdaExpressionsCleanUpCore(options), IProposalRelevance.CONVERT_TO_ANONYMOUS_CLASS_CREATION, context, CodeActionKind.Refactor);
+		resultingCollections.add(proposal);
+		return true;
 	}
 
 }
